requirements.txt
langchain
sentence_transformers
torch
sentencepiece
transformers
accelerate
pypdf
tiktoken
streamlit
chromadb
base64
nltk
virtualenv